api: serialize per-user sync execution and reorganize state modules
- Move state helpers into api.state.status_services module.
- Introduce sync runner with per-user asyncio locks.
- Prevent concurrent sync execution for same user.
- Replace direct sync_lastfm_api calls with an async wrapper.
