+++ apps/api/routers/websockets.py
@@ -7,7 +7,7 @@ from memoryfm.models.sync_status import (
     UserExist,
 )
 
-from api.state import get_ensure_user_status, get_sync_status
+from api.state.status_services import get_ensure_user_status, get_sync_status
 
 router = APIRouter()
